requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.10
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
exceptiongroup==1.2.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
iniconfig==2.0.0
jwt_auth==0.4.3
packaging==23.2
pluggy==1.4.0
promise==2.3
PyJWT==1.7.1
pytest==8.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
